Modify the code to call the lastround function.
For some reason, this functino was implemented but unused,
and its implementation was inlined instead.  However,
the shift and substitute functions were not declared ahead
of time, which makes them into variable arguments functions,
which leads to complications with SAW.

To work around this, we just call the function.
Alternatively, we could simply declare the types of the
substitution and shifting functions.